fix(user): 토큰 재발급되도록 수정

+++ apps/user/src/apis/token/refresh.ts
@@ -4,18 +4,21 @@ import { Cookie } from '../cookie/cookie';
 import { Storage } from '../storage/storage';
 
 const refreshToken = async () => {
-  try {
-    const { data } = await maru.patch('/auth', null, {
-      headers: {
-        'Refresh-Token': `${Cookie.getItem(TOKEN.REFRESH)}`,
-      },
-    });
-    Storage.setItem(TOKEN.ACCESS, data.data.accessToken);
-  } catch {
+  const refreshTokenValue = Cookie.getItem(TOKEN.REFRESH);
+
+  if (!refreshTokenValue) {
+    alert('다시 로그인 해주세요.');
     window.location.href = ROUTES.LOGIN;
-    alert('다시 로그인 해주세요');
     localStorage.clear();
+    return;
   }
+
+  const { data } = await maru.patch('/auth', null, {
+    headers: {
+      'Refresh-Token': refreshTokenValue,
+    },
+  });
+  Storage.setItem(TOKEN.ACCESS, data.data.accessToken);
 };
 
 export default refreshToken;
